Adding Executable description class (#454)
* Adding Executable description class

Part of implementation of refactor for #114

Distro A; OPSEC #2893


* Cleaned up access to substituted values

Modified handling of cmd parameter as described in #263

Distro A; OPSEC #2893


* Initial implementation of execute_local

Distro A; OPSEC #4584


* Fixed minor bugs to verify unit tests

Distro A; OPSEC #4584


* Unit test fixes

Distro A; OPSEC #4584


* Adjust for launch_ros modifications, add unit tests

Distro A; OPSEC #4584


* Apply minor updates to address feedback

Distro A; OPSEC #4584


* Refactor arguments to apply to executable

Distro A; OPSEC #4584


* Fix order of parameters

Distro A; OPSEC #4584


* Fixed default log name prefix

Distro A; OPSEC #4584


* Fix unit tests

Distro A; OPSEC #4584


* Removed some language from file headers

Distro A, OPSEC #4584. You may have additional rights; please see https://rosmilitary.org/faq/?category=ros-2-license


* Add prefix filtering to Executable description class.

Distro A, OPSEC #4584. You may have additional rights; please see https://rosmilitary.org/faq/?category=ros-2-license


* Formatting fixes.

Distro A, OPSEC #4584. You may have additional rights; please see https://rosmilitary.org/faq/?category=ros-2-license


* Fixed indentation in descriptions/executable.py.

Distro A, OPSEC #4584. You may have additional rights; please see https://rosmilitary.org/faq/?category=ros-2-license


* Fixed namechange missed during rebase.

Distro A, OPSEC #4584. You may have additional rights; please see https://rosmilitary.org/faq/?category=ros-2-license
